refactor: move back to `core`
